Create task-helpers folder
Rails is currently setup to autoload all ruby files in the lib/
folder. However, we also want a place to store code to be used by rake
tasks that don't need to be loaded by rails. Those classes can live in
the task-helpers folder and be required 'manually' in the rake tasks.