[#238] add pagination links
This uses's Kiminari.paginate_array with a custom count.

The supplied count -- taggings.count * 2 -- is technically an estimate.
Most tagables will have two edit rows in the table,
one row for when they were tagged and another row for
that tagable's most recent update. However, in the case that
the tagable's most recent update WAS the tagging event (.tag), then there will
only be only one row in the table for that tagable. Meaning the
supplied count is likely a slight overcount.

The effect is mostly inconsequential except for those brave users who
want to click 'next page' until the very end. The last page might be
missing or truncated.