[#357] Validate username and wrap user models in a transaction
If the SfGuardUserProfile had an invalid field, the error message
wouldn't appear in devise user messages, causing the account
sign up to fail without providing any visual indication to the user.

Errors message are now handled in the controller, and the mdoels are saved in a
transaction to ensure that SfGuardUserProfiles are never persisted
without the user model also being saved.