[#371] use different param names for entities tab
If provided a subtype as happens with entities -- Org and Person --
the tagable list will use a differently named page parameter.
For example: 'person_page'. If there are no subtypes, the default
page param, 'page', is used.

Additionally, invocation syntax changed to use keywords arguments
in TagsController for all show and edits actions.