[#483] add field "elected_term" to Membership
This column is serilized as a OpenStruct and will contain
extra information particular to relationships for terms for elected
office.

Initially, it will be used for federal legislators with the
following fields:
  * type (congress)
  * party
  * state
  * district

However, perhaps when state legislator are importer we might consider
storing information for state legislator in this field as well.