Install select2 for dynamic dropdowns
Note: this requires a workaround to deal with the current dual use of both webpacker and sprockets, since we need to use select2 in basic DOM scripts which are currently managed by sprockets, but it makes sense to manage it via yarn, which currently only manages packages accessible to webpacker. Here this is handled with a couple of simple symlinks from vendor/assets/javascripts and vendor/assets/stylesheets.